NTP now works!
I've changed to using the standard port, and moved out of test mode. The network
code now assumes there's a password. I've also added a pretty-printer for times.
All times are currently UTC, with no concept of local time so far. That will be
important for the upcoming calendar work.