Added a calendar, and local time support.
I had considered several designs for this calendar system, and finally
decided on a hard-coded, compile time calendar approach. This has some
pitfalls, such as requiring sorting at input time, but for an embedded
project with specialized needs and customized deployment, that's not a
very large stretch. This system will need redeployment with the change
in Daylight Savings time, but as is, the flexible calendar offers lots
of space to have multiple classes of events driven by a single switch,
or other more complex use cases than just turning the lava lamp that's
in my office on and off.